In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,8 +8,8 @@
 int main(void)
 {
 
-long long i, num1 = 1, num2 = 2;
-long long temp, n = 50;
+long i, num1 = 1, num2 = 2;
+long  temp, n = 50;
 for (i = 0; i <= n; i++)
 {
 printf("%lld, ", num1);
